requirements.txt
Jinja2~=3.1.2
jsonschema~=4.17.3
beautifulsoup4~=4.11.1
markdown-it-py[linkify]~=2.1.0
typer[all]~=0.7.0
